chore(gsd): delete dead planner-handoff module
planner-handoff.ts has had zero production references since 2a6867d9c
removed the planner command surface: nothing imports it, it has no
manifest or catalog entry, and /gsd planner falls through to
unknown-command handling (#2141). The planner-relaunch branch that would
have rewired it never landed, so the module and its test are deleted.
The regression pins that planner stays absent from the /gsd surface move
into catalog, autocomplete-regressions, and dispatch-rule-coverage
tests; the fall-through pin now drives the real handleGSDCommand
dispatcher instead of a single handler.
