+++ src/resources/extensions/gsd/tests/autocomplete-regressions-1675.test.ts
@@ -121,6 +121,17 @@ test("bare /gsd skip shows usage and does not fall through to unknown-command wa
   );
 });
 
+test("/gsd planner falls through to unknown-command handling", async () => {
+  const ctx = createMockCtx();
+
+  await handleGSDCommand("planner", ctx as any, {} as any);
+
+  assert.ok(
+    ctx.notifications.some((n) => n.message.startsWith("Unknown: /gsd planner")),
+    "planner should be reported as an unknown /gsd command",
+  );
+});
+
 test("direct loop verbs do not fall through to unknown-command warning", async () => {
   const loopVerbs = [
     "research-milestone",
